backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend for dream team project"
readme = "README.md"
requires-python = ">=3.10, <3.13"
dependencies = [
    "applicationinsights==0.11.10",
    "azure-identity==1.19.0",
    "python-dotenv>=1.1.0",
    "autogen-agentchat==0.5.7",
    "autogen-ext[azure,docker,mcp]==0.5.7",
    "autogen-core==0.5.7",
    "azure-search-documents==11.6.0b10",
    "markitdown==0.0.2",
    "aiofiles==24.1.0",
    "aiohttp==3.13.3",
    "playwright==1.52.0",
    "openai==1.66.0",
    "azure-storage-blob==12.25.0",
    "fastapi-cors==0.0.6",
    "fastapi==0.115.12",
    "uvicorn==0.34.0",
    "python-multipart==0.0.20",
    "tiktoken==0.9.0",
    "jinja2==3.1.6",
    "azure-cosmos==4.9.0",
    "fastmcp==2.13.0",
    "mcp==1.23.0",
    "azure-communication-email==1.0.0"
]
